+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/service/DaoOnlyCrudService.java
@@ -75,12 +75,10 @@ public class DaoOnlyCrudService<K extends Key, E extends Entity<K>> implements C
     }
 
     private K internalInsert(E element) throws Exception {
-        if (internalExists(element.getKey())) {
-            throw new ServiceException(ServiceExceptionCodes.ENTITY_EXISTED);
-        }
-
         if (Objects.isNull(element.getKey())) {
             element.setKey(keyFetcher.fetchKey());
+        } else if (internalExists(element.getKey())) {
+            throw new ServiceException(ServiceExceptionCodes.ENTITY_EXISTED);
         }
         return dao.insert(element);
     }
@@ -161,6 +159,7 @@ public class DaoOnlyCrudService<K extends Key, E extends Entity<K>> implements C
         }
     }
 
+    @SuppressWarnings("DuplicatedCode")
     @Override
     public K insertOrUpdate(E element) throws ServiceException {
         try {